+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.5] - 2020-03-19
+
+- Update Instrument `notes` and allow multi trigger of same note
+- Update Rollup config to includeDependencies
+- Remove `tone` as peerDep and leave as dep
+
 ## [0.3.4] - 2020-03-16
 
 - Add `fast-deep-equal` to improve steps update performance on `Track`